Fix browser compatibility issues in shared modules
