+++ shared/node-build-utils.ts
@@ -1,0 +1,23 @@
+import { copyFileSync } from 'fs';
+import { resolve } from 'path';
+
+/**
+ * Creates a Vite plugin that copies files after build
+ * @param sourceDir - Source directory
+ * @param targetDir - Target directory
+ * @param files - Array of files to copy
+ */
+export function createCopyFilesPlugin(sourceDir: string, targetDir: string, files: string[]) {
+  return {
+    name: 'copy-files',
+    closeBundle: () => {
+      for (const file of files) {
+        try {
+          copyFileSync(resolve(sourceDir, file), resolve(targetDir, file));
+        } catch (error) {
+          console.warn(`Warning: Could not copy ${file}: ${error}`);
+        }
+      }
+    }
+  };
+}
